Rename handwritten_digit_utils.py to extraction_utilities.py
This file will consolidate functions and constants that would otherwise be duplicated between files.
